+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-  - Add `v7 list-reusable-models` command.
+  - Add `v7 list-reusable-models` command. [#146](https://github.com/nablaflow/cli/pull/146)
 
 # 1.1.0 - 2026-01-28