+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.3.0 - 2026-06-04
 
 ## AeroCloud